# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "distribute MCP Service",
    "description": "MCP (Model Context Protocol) server for distribute. Exposes campaign management tools via MCP protocol over HTTP.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "https://mcp.distribute.you"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "distribute org API key (Authorization: Bearer <key>)"
      }
    },
...
  },
  "paths": {
...
    "/mcp": {
      "post": {
        "tags": [
          "MCP"
        ],
        "summary": "MCP JSON-RPC endpoint",
        "description": "Handles MCP protocol requests. Use mcp-session-id header for session management.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "mcp-session-id",
            "schema": {
              "type": "string"
            },
            "description": "Session ID (returned after first request)"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "jsonrpc": {
                    "type": "string",
                    "example": "2.0"
                  },
                  "method": {
                    "type": "string",
                    "example": "tools/call"
                  },
                  "params": {
                    "type": "object"
                  },
                  "id": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JSON-RPC response"
          }
        }
      },
      "get": {
        "tags": [
          "MCP"
        ],
        "summary": "MCP SSE endpoint for server notifications",
        "parameters": [
          {
            "in": "header",
            "name": "mcp-session-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream"
          }
        }
      },
      "delete": {
        "tags": [
          "MCP"
        ],
        "summary": "Close MCP session",
        "parameters": [
          {
            "in": "header",
            "name": "mcp-session-id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session closed"
          }
        }
      }
    },
...
  }
}